Remove pre-delete use pre-get and delflg instead
If a professor try to delete a course not owned by him/her.
pre-delete will delete all students of the course

so, we need to pre-get stulist or it will disappear after the course be
deleted.

then use delflg (deletion flag) to ensure that the professor is actually
deleting his/hers own course.

check if delflg is not zero, if it is, delete the related records in
Course_taken and add records into table Course_change.